requirements.txt
fastapi
uvicorn
pyswisseph
pytz